pyproject.toml
[project]
name = "post-training-of-deepseek-ocr"
version = "0.1.0"
description = "Applying post-training techniques to improve DeepSeek OCR model performance"
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "addict>=2.4.0",
    "datasets>=4.4.1",
    "easydict>=1.13",
    "einops>=0.8.1",
    "hf-xet>=1.2.0",
    "img2pdf>=0.6.3",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "notebook>=7.5.0",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "psutil>=7.1.3",
    "pymupdf>=1.26.6",
    "tokenizers==0.20.3",
    "transformers==4.46.3",
    "torch>=2.6.0",
    "torchvision>=0.24.1",
    "trl==0.12.2",
    "peft==0.13.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }
